Split and refactor Chat Window components
